Fix bugs found by static analyzer
- SFTPController.m: fix promisedNamesFromPlists using uninitialized
  local `plist` before first assignment on lines 4788/4792; correct
  to `plists` (the parameter)
- SFTPSplitView.m: change dividerThickness return type float → CGFloat
  to match NSSplitView override signature on 64-bit
- SFTPController.m, SSHTunnel.m, SCPController.m: restructure -init
  to call [super init] before accessing instance variables

+++ SFTPSplitView.m
@@ -7,7 +7,7 @@
 
 @implementation SFTPSplitView
 
-- ( float )dividerThickness
+- ( CGFloat )dividerThickness
 {
     return( 20.0 );
 }
